requirements.txt
# Core dependencies
pandas==2.2.3
numpy==1.26.0
PyYAML==6.0.1
python-dotenv==1.0.0
click==8.1.7
pydantic==2.3.0

# Machine Learning & RL
# PyTorch - choose one of these based on your system:
#--extra-index-url https://download.pytorch.org/whl/cu118  # For NVIDIA GPU
#torch==2.0.1+cu118  # For NVIDIA GPU
torch==2.2.2  # For CPU or Apple Silicon

gymnasium==0.29.1  # New name for gym
stable-baselines3==2.1.0
scikit-learn==1.3.0

# Technical Analysis
#TA-Lib
pandas-ta==0.3.14b
TA-Lib
# Trading & Exchange
python-binance>=1.0.19
ccxt>=4.0.0
aiohttp>=3.8.0
websockets>=10.0
ujson>=5.0.0

# Database
aiosqlite==0.19.0
asyncpg==0.28.0

# Monitoring & Logging
wandb
tensorboard==2.14.0
matplotlib==3.7.2
seaborn==0.12.2
cryptography==41.0.3

# Development
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-timeout==2.3.1
black==23.7.0
flake8==6.1.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# pyfolio

